advanced-api-project/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-filter==25.1
asgiref==3.9.2
sqlparse==0.5.3